+++ addons/assets/script_classes_vehicles_cup.hpp
@@ -1,0 +1,22 @@
+// Should only be included in CUP subconfigs.
+
+// Turrets - CUP
+class TACU_Assets_AGS_base;
+class TACU_Assets_D30_base;
+class TACU_Assets_D30AT_base;
+class TACU_Assets_DSHKM_base;
+class TACU_Assets_DSHKM_Mini_base;
+class TACU_Assets_FIM92_base;
+class TACU_Assets_IGLA_base;
+class TACU_Assets_KORD_base;
+class TACU_Assets_KORD_Mini_base;
+class TACU_Assets_KORNET_AT4_base;
+class TACU_Assets_L11A1_base;
+class TACU_Assets_L11A1_MiniTripod_base;
+class TACU_Assets_L134A1_base;
+class TACU_Assets_M119_base;
+class TACU_Assets_M220_TOW_base;
+class TACU_Assets_Metis_base;
+class TACU_Assets_MK19_base;
+class TACU_Assets_RBS70_base;
+class TACU_Assets_SearchLight_base;
