+++ addons/assets/subconfig_cup/CfgVehicles.hpp
