+++ addons/assets/subconfig_cup/script_component.hpp
@@ -1,0 +1,5 @@
+#define COMPONENT assets_cup
+#define COMPONENT_BEAUTIFIED Assets CUP
+#include "\x\tacu\addons\main\script_mod.hpp"
+#include "\x\tacu\addons\main\script_macros.hpp"
+#include "..\script_macros.hpp"
